refactor: data moved to different file
