+++ package.json
@@ -9,10 +9,10 @@
   "private": false,
   "scripts": {
     "task:1": "yarn run nodemon ./src/task_1.ts",
-    "task:2": "yarn run nodemon ./src/task_2.ts",
+    "task:2": "yarn run nodemon ./src/task_2/index.ts",
     "build": "yarn run rimraf ./build && yarn run tsc",
     "start:1": "node ./build/task_1.js",
-    "start:2": "node ./build/task_2.js"
+    "start:2": "node ./build/task_2/index.js"
   },
   "devDependencies": {
     "@types/node": "^20.12.7",
