+++ src/task_2/data.ts
@@ -1,4 +1,4 @@
-const posts = [
+export const posts = [
   {
     id: '62e69d5a5458aac0ed320b35',
     title: 'id labore ex et quam laborum',
@@ -35,42 +35,3 @@ const posts = [
     body: 'maiores sed dolores similique labore et inventore etquasi temporibus esse sunt id eteos voluptatem aliquamratione corporis molestiae mollitia quia et magnam dolor',
   },
 ]
-
-type DataItemT = {
-  id: string
-  title: string
-  body: string
-}
-
-type UnnormalizedDataT = Array<DataItemT>
-
-type NormalizedDataT = {
-  byId: { [key: string]: DataItemT }
-  allIds: string[]
-}
-
-const normalizeData = (
-  unnormalizedData: UnnormalizedDataT
-): NormalizedDataT => {
-  const normalizedData = unnormalizedData.reduce(
-    (acc: NormalizedDataT, item: DataItemT): NormalizedDataT => {
-      acc.byId[item.id] = item
-      acc.allIds.push(item.id)
-      return acc
-    },
-    { byId: {}, allIds: [] }
-  )
-  return normalizedData
-}
-
-console.log(normalizeData(posts))
-/**
- * {
- *    byId: {
- *      62e69d5a5458aac0ed320b35: { id: '...', title: '...', body: '...' },
- *      62e69d5a5458aac0ed320b1c: { id: '...', title: '...', body: '...' },
- *      ...
- *    },
- *    allIds: ['62e69d5a5458aac0ed320b35', '62e69d5a5458aac0ed320b1c', ...]
- * }
- */
